# swagger.yml
openapi: 3.0.3
info:
  title: EduQuest
  description: |
    Приложение для улучшения качества образования с элементами геймификации.
    **Основные возможности:**
    - Анонимная обратная связь после пар
    - Вопросы во время занятий
    - Викторины с начислением баллов
    - Карточки для повторения
    - Тренажёр решения задач
    - Возможность обмена баллов на награды
    **Роли пользователей:**
    - student — студент
    - teacher — преподаватель
    - admin — администратор
  version: 1.0.0
  contact:
    name: EduQuest Team
    email: support@eduquest.ru
servers:
  - url: http://localhost/eduquest
    description: Локальный сервер разработки
  - url: https://api.eduquest.ru
    description: Продакшен сервер
...
paths:
...
  /api/trainer/problems:
    post:
      tags: [Trainer]
      summary: Создать новую задачу для тренажёра
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Problem'
      responses:
        '201':
          description: Задача создана
        '400':
          description: Некорректные данные
        '403':
          description: Только преподаватели могут создавать задачи
        '401':
          description: Требуется аутентификация
        '500':
          description: Ошибка сервера
    get:
      tags: [Trainer]
      summary: Получить задачи для тренажёра
      security:
        - BearerAuth: []
      parameters:
        - name: difficulty
          in: query
          schema:
            type: string
            enum: [easy, medium, hard]
        - name: subject
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Список задач
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Problem'
        '401':
          description: Требуется аутентификация
        '500':
          description: Ошибка сервера
  /api/trainer/problems/{id}/solve:
    post:
      tags: [Trainer]
      summary: Решить задачу
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Solution'
      responses:
        '200':
          description: Результат решения
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SolutionResult'
        '400':
          description: Некорректные данные
        '401':
          description: Требуется аутентификация
        '500':
          description: Ошибка сервера

  # Награды
  /api/rewards:
    post:
      tags: [Rewards]
      summary: Создать новую награду
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Reward'
      responses:
        '201':
          description: Награда создана
        '400':
          description: Некорректные данные
        '403':
          description: Только администраторы могут создавать награды
        '401':
          description: Требуется аутентификация
        '500':
          description: Ошибка сервера
    get:
      tags: [Rewards]
      summary: Получить доступные награды
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Список наград
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Reward'
        '401':
          description: Требуется аутентификация
        '500':
          description: Ошибка сервера

  # Викторины
  /api/quiz/questions:
    post:
      tags: [Quiz]
      summary: Создать новый вопрос для викторины
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuizQuestion'
      responses:
        '201':
          description: Вопрос викторины создан
        '400':
          description: Некорректные данные
        '403':
          description: Только преподаватели могут создавать вопросы викторины
        '401':
          description: Требуется аутентификация
        '500':
          description: Ошибка сервера
  /api/quiz:
    get:
      tags: [Quiz]
      summary: Получить вопросы викторины
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Вопросы викторины
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/QuizQuestion'
        '401':
          description: Требуется аутентификация
        '500':
          description: Ошибка сервера
  /api/quiz/submit:
    post:
      tags: [Quiz]
      summary: Отправить ответы викторины
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/QuizAnswer'
      responses:
        '200':
          description: Результаты викторины
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuizResult'
        '400':
          description: Некорректные данные
        '401':
          description: Требуется аутентификация
        '500':
          description: Ошибка сервера

  # Аутентификация
...
  /api/users/{id}/points:
    get:
      tags: [Users, Quiz]
      summary: Получить баллы пользователя
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            example: 1
      responses:
        '200':
          description: Количество баллов
          content:
            application/json:
              schema:
                type: integer
                example: 150
        '401':
          description: Требуется аутентификация
        '403':
          description: Доступ запрещён
        '404':
          description: Пользователь не найден
        '500':
          description: Ошибка сервера
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          example: 1
        first_name:
          type: string
          example: "Иван"
        last_name:
          type: string
          example: "Иванов"
        email:
          type: string
          format: email
          example: "ivan@university.edu"
        role:
          type: string
          enum: [student, teacher, admin]
          example: "student"
        points:
          type: integer
          example: 150
        group:
          $ref: '#/components/schemas/Group'
        department:
          type: string
          example: "Математический факультет"
      required: [id, first_name, last_name, email, role, group]
    Group:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: "МАТ-101"
        faculty:
          type: string
          example: "Математический факультет"
      required: [id, name]
...
    QuizQuestion:
      type: object
      properties:
        id:
          type: integer
          example: 1
        text:
          type: string
          example: "Что такое производная функции?"
        options:
          type: array
          items:
            type: string
          example: ["Предел", "Интеграл", "Дифференциал", "Матрица"]
        correct_answer:
          type: integer
          example: 0
        points:
          type: integer
          example: 10
      required: [text, options]
    QuizAnswer:
      type: object
      properties:
        question_id:
          type: integer
          example: 1
        selected_option_index:
          type: integer
          example: 0
      required: [question_id, selected_option_index]
    QuizResult:
      type: object
      properties:
        total_points:
          type: integer
          example: 30
        correct_answers:
          type: integer
          example: 3
        total_questions:
          type: integer
          example: 5
        earned_points:
          type: integer
          example: 10
...
    Problem:
      type: object
      properties:
        id:
          type: integer
          example: 1
        text:
          type: string
          example: "Найдите производную функции f(x) = x^2 + 3x + 1"
        difficulty:
          type: string
          enum: [easy, medium, hard]
          example: "easy"
        subject:
          type: string
          example: "Математика"
        solution_steps:
          type: array
          items:
            type: string
          example: ["f'(x) = 2x + 3", "Ответ: 2x + 3"]
      required: [text, difficulty]
    Solution:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/User'
        answer:
          type: string
          example: "2x + 3"
      required: [user, answer]
    SolutionResult:
      type: object
      properties:
        is_correct:
          type: boolean
          example: true
        explanation:
          type: string
          example: "Правильно! Производная x^2 равна 2x, производная 3x равна 3, производная константы 0"
        points_earned:
          type: integer
          example: 5
    Reward:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: "Футболка с логотипом факультета"
        description:
          type: string
          example: "Хлопковая футболка с принтом"
        cost:
          type: integer
          example: 100
        category:
          type: string
          example: "merch"
      required: [name, cost]